guessDist {gbm}	R Documentation

gbm internal functions

Description

Helper functions for preprocessing data prior to building a "gbm" object.

Usage

guessDist(y)

getCVgroup(distribution, class.stratify.cv, y, i.train, cv.folds, group)

getStratify(strat, d)

checkMissing(x, y)

checkWeights(w, n)

checkID(id)

checkOffset(o, y)

getVarNames(x)

gbmCluster(n)

Arguments

y
The response variable.

class.stratify.cv
Whether or not to stratify, if provided by the user.

i.train
Computed internally by gbm.

cv.folds
The number of cross-validation folds.

group
The group, if using distibution = "pairwise".

strat
Whether or not to stratify.

d, distribution
The distribution, either specified by the user or implied.

x
The design matrix.

w
The weights.

n
The number of cores to use in the cluster.

id
The interaction depth.

o
The offset.

Details

These are functions used internally by gbm and not intended for direct use by the user.

[Package gbm version 2.2.3 Index]