predict.gbm {gbm}	R Documentation

Predict method for GBM Model Fits

Description

Predicted values based on a generalized boosted model object

Usage

## S3 method for class 'gbm'
predict(object, newdata, n.trees, type = "link", single.tree = FALSE, ...)

Arguments

object
Object of class inheriting from (gbm.object)

newdata
Data frame of observations for which to make predictions

n.trees
Number of trees used in the prediction. n.trees may be a vector in which case predictions are returned for each iteration specified

type
The scale on which gbm makes the predictions

single.tree
If single.tree=TRUE then predict.gbm returns only the predictions from tree(s) n.trees

...
further arguments passed to or from other methods

Details

predict.gbm produces predicted values for each observation in newdata using the the first n.trees iterations of the boosting sequence. If n.trees is a vector than the result is a matrix with each column representing the predictions from gbm models with n.trees[1] iterations, n.trees[2] iterations, and so on.

The predictions from gbm do not include the offset term. The user may add the value of the offset to the predicted value if desired.

If object was fit using gbm.fit there will be no Terms component. Therefore, the user has greater responsibility to make sure that newdata is of the same format (order and number of variables) as the one originally used to fit the model.

Value

Returns a vector of predictions. By default the predictions are on the scale of f(x). For example, for the Bernoulli loss the returned value is on the log odds scale, poisson loss on the log scale, and coxph is on the log hazard scale.

If type="response" then gbm converts back to the same scale as the outcome. Currently the only effect this will have is returning probabilities for bernoulli and expected counts for poisson. For the other distributions "response" and "link" return the same.

Author(s)

Greg Ridgeway gregridgeway@gmail.com

See Also

gbm, gbm.object

[Package gbm version 2.2.3 Index]