quantile.rug {gbm}	R Documentation

Quantile rug plot

Description

Marks the quantiles on the axes of the current plot.

Usage

## S3 method for class 'rug'
quantile(x, prob = 0:10/10, ...)

Arguments

x
A numeric vector.

prob
The quantiles of x to mark on the x-axis.

...
Additional optional arguments to be passed onto rug

Value

No return values.

Author(s)

Greg Ridgeway gregridgeway@gmail.com.

See Also

plot, quantile, jitter, rug.

Examples

x <- rnorm(100)
y <- rnorm(100)
plot(x, y)
quantile.rug(x)

[Package gbm version 2.2.3 Index]