L1 shrinkage of the predictor variables in a GBM

Performs recursive shrinkage in each of the trees in a GBM fit using different shrinkage parameters for each variable.

shrink.gbm(object, n.trees, lambda = rep(10, length(object$var.names)),
  ...)

Arguments

object
A gbm.object.

n.trees
Integer specifying the number of trees to use.

lambda
Vector of length equal to the number of variables containing the shrinkage parameter for each variable.

...
Additional optional arguments. (Currently ignored.)

Value

predF

Predicted values from the shrunken tree

objective

The value of the loss function associated with the predicted values

gradient

A vector with length equal to the number of variables containing the derivative of the objective function with respect to beta, the logit transform of the shrinkage parameter for each variable

Details

This function is currently experimental. Used in conjunction with a gradient ascent search for inclusion of variables.

Note

Warning: This function is experimental.

References

Hastie, T. J., and Pregibon, D. https://web.stanford.edu/~hastie/Papers/shrink_tree.pdf. AT&T Bell Laboratories Technical Report (March 1990).

See also