shrink.gbm.pred {gbm}	R Documentation

Predictions from a shrunked GBM

Description

Makes predictions from a shrunken GBM model.

Usage

shrink.gbm.pred(object, 
                newdata, 
                n.trees, 
                lambda = rep(1, length(object$var.names)), 
                ...)

Arguments

object
a gbm.object

newdata
dataset for predictions

n.trees
the number of trees to use

lambda
a vector with length equal to the number of variables containing the shrinkage parameter for each variable

...
other parameters (ignored)

Value

A vector with length equal to the number of observations in newdata containing the predictions

Warning

This function is experimental

Author(s)

Greg Ridgeway gregridgeway@gmail.com

See Also

shrink.gbm, gbm

[Package gbm version 2.1.1 Index]