gbm-package {gbm}	R Documentation

Generalized Boosted Regression Models

Description

This package implements extensions to Freund and Schapire's AdaBoost algorithm and J. Friedman's gradient boosting machine. Includes regression methods for least squares, absolute loss, logistic, Poisson, Cox proportional hazards partial likelihood, multinomial, t-distribution, AdaBoost exponential loss, Learning to Rank, and Huberized hinge loss.

Details

Package: 	 gbm

Version: 	 2.1

Date: 	 2013-05-10

Depends: 	 R (>= 2.9.0), survival, lattice, mgcv

License: 	 GPL (version 2 or newer)

URL: 	 http://code.google.com/p/gradientboostedmodels/

Index:

basehaz.gbm             Baseline hazard function
calibrate.plot          Calibration plot
gbm                     Generalized Boosted Regression Modeling
gbm.object              Generalized Boosted Regression Model Object
gbm.perf                GBM performance
plot.gbm                Marginal plots of fitted gbm objects
predict.gbm             Predict method for GBM Model Fits
pretty.gbm.tree         Print gbm tree components
quantile.rug            Quantile rug plot
relative.influence      Methods for estimating relative influence
shrink.gbm              L1 shrinkage of the predictor variables in a GBM
shrink.gbm.pred         Predictions from a shrunked GBM
summary.gbm             Summary of a gbm object

Further information is available in the following vignettes:

gbm 	 Generalized Boosted Models: A guide to the gbm package (source, pdf)

Author(s)

Greg Ridgeway gregridgeway@gmail.com with contributions by Daniel Edwards, Brian Kriegler, Stefan Schroedl and Harry Southworth.

References

Y. Freund and R.E. Schapire (1997) “A decision-theoretic generalization of on-line learning and an application to boosting,” Journal of Computer and System Sciences, 55(1):119-139.

G. Ridgeway (1999). “The state of boosting,” Computing Science and Statistics 31:172-181.

J.H. Friedman, T. Hastie, R. Tibshirani (2000). “Additive Logistic Regression: a Statistical View of Boosting,” Annals of Statistics 28(2):337-374.

J.H. Friedman (2001). “Greedy Function Approximation: A Gradient Boosting Machine,” Annals of Statistics 29(5):1189-1232.

J.H. Friedman (2002). “Stochastic Gradient Boosting,” Computational Statistics and Data Analysis 38(4):367-378.

The MART website.

[Package gbm version 2.1.1 Index]