Generalized Boosted Regression Model Object

These are objects representing fitted gbms.

Value

initF

the "intercept" term, the initial predicted value to which trees make adjustments

fit

a vector containing the fitted values on the scale of regression function (e.g. log-odds scale for bernoulli, log scale for poisson)

train.error

a vector of length equal to the number of fitted trees containing the value of the loss function for each boosting iteration evaluated on the training data

valid.error

a vector of length equal to the number of fitted trees containing the value of the loss function for each boosting iteration evaluated on the validation data

cv.error

if cv.folds<2 this component is NULL. Otherwise, this component is a vector of length equal to the number of fitted trees containing a cross-validated estimate of the loss function for each boosting iteration

oobag.improve

a vector of length equal to the number of fitted trees containing an out-of-bag estimate of the marginal reduction in the expected value of the loss function. The out-of-bag estimate uses only the training data and is useful for estimating the optimal number of boosting iterations. See gbm.perf

trees

a list containing the tree structures. The components are best viewed using pretty.gbm.tree

c.splits

a list of all the categorical splits in the collection of trees. If the trees[[i]] component of a gbm object describes a categorical split then the splitting value will refer to a component of c.splits. That component of c.splits will be a vector of length equal to the number of levels in the categorical split variable. -1 indicates left, +1 indicates right, and 0 indicates that the level was not present in the training data

cv.fitted

If cross-validation was performed, the cross-validation predicted values on the scale of the linear predictor. That is, the fitted values from the ith CV-fold, for the model having been trained on the data in all other folds.

Structure

The following components must be included in a legitimate gbm object.

See also