plot.gbm {gbm}	R Documentation

Marginal plots of fitted gbm objects

Description

Plots the marginal effect of the selected variables by "integrating" out the other variables.

Usage

## S3 method for class 'gbm'
plot(x,
     i.var = 1,
     n.trees = x$n.trees,
     continuous.resolution = 100,
     return.grid = FALSE,
     type = "link",
     ...)

Arguments

x
a gbm.object fitted using a call to gbm

i.var
a vector of indices or the names of the variables to plot. If using indices, the variables are indexed in the same order that they appear in the initial gbm formula. If length(i.var) is between 1 and 3 then plot.gbm produces the plots. Otherwise, plot.gbm returns only the grid of evaluation points and their average predictions

n.trees
the number of trees used to generate the plot. Only the first n.trees trees will be used

continuous.resolution
The number of equally space points at which to evaluate continuous predictors

return.grid
if TRUE then plot.gbm produces no graphics and only returns the grid of evaluation points and their average predictions. This is useful for customizing the graphics for special variable types or for dimensions greater than 3

type
the type of prediction to plot on the vertical axis. See predict.gbm

...
other arguments passed to the plot function

Details

plot.gbm produces low dimensional projections of the gbm.object by integrating out the variables not included in the i.var argument. The function selects a grid of points and uses the weighted tree traversal method described in Friedman (2001) to do the integration. Based on the variable types included in the projection, plot.gbm selects an appropriate display choosing amongst line plots, contour plots, and lattice plots. If the default graphics are not sufficient the user may set return.grid=TRUE, store the result of the function, and develop another graphic display more appropriate to the particular example.

Value

Nothing unless return.grid is true then plot.gbm produces no graphics and only returns the grid of evaluation points and their average predictions.

Author(s)

Greg Ridgeway gregridgeway@gmail.com

References

J.H. Friedman (2001). "Greedy Function Approximation: A Gradient Boosting Machine," Annals of Statistics 29(4).

See Also

gbm, gbm.object, plot

[Package gbm version 2.1.1 Index]