quantile.rug {gbm}	R Documentation

Quantile rug plot

Description

Marks the quantiles on the axes of the current plot.

Usage

quantile.rug(x,prob=(0:10)/10,...)

Arguments

x
a numeric vector.

prob
the quantiles of x to mark on the x-axis.

...
additional graphics parameters currently ignored.

Value

No return values

Author(s)

Greg Ridgeway gregridgeway@gmail.com

See Also

plot, quantile, jitter, rug.

Examples

x <- rnorm(100)
y <- rnorm(100)
plot(x,y)
quantile.rug(x)

[Package gbm version 2.1.1 Index]