reconstructGBMdata {gbm}	R Documentation

Reconstruct a GBM's Source Data

Description

Helper function to reconstitute the data for plots and summaries. This function is not intended for the user to call directly.

Usage

reconstructGBMdata(x)

Arguments

x
a gbm.object initially fit using gbm

Value

Returns a data used to fit the gbm in a format that can subsequently be used for plots and summaries

Author(s)

Harry Southworth

See Also

gbm, gbm.object

[Package gbm version 2.1.1 Index]