shrink.gbm {gbm}	R Documentation

L1 shrinkage of the predictor variables in a GBM

Description

Performs recursive shrinkage in each of the trees in a GBM fit using different shrinkage parameters for each variable.

Usage

shrink.gbm(object, 
           n.trees, 
           lambda = rep(10, length(object$var.names)), 
           ...)

Arguments

object
A gbm.object

n.trees
the number of trees to use

lambda
a vector with length equal to the number of variables containing the shrinkage parameter for each variable

...
other parameters (ignored)

Details

This function is currently experimental. Used in conjunction with a gradient ascent search for inclusion of variables.

Value

predF
Predicted values from the shrunken tree

objective
The value of the loss function associated with the predicted values

gradient
A vector with length equal to the number of variables containing the derivative of the objective function with respect to beta, the logit transform of the shrinkage parameter for each variable

Warning

This function is experimental.

Author(s)

Greg Ridgeway gregridgeway@gmail.com

References

Hastie, T. J., and Pregibon, D. Shrinking Trees. AT&T Bell Laboratories Technical Report (March 1990).

See Also

shrink.gbm.pred, gbm

[Package gbm version 2.1.1 Index]